rename(network-api): 시스템 검증 패키지 이름 변경
