+++ modules/network-api/src/main/java/com/whoz_in/network_api/system/CommandsInstalledValidator.java
@@ -1,4 +1,4 @@
-package com.whoz_in.network_api.system_validator;
+package com.whoz_in.network_api.system;
 
 
 import com.whoz_in.network_api.common.process.TransientProcess;
